backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
dj-database-url>=2.1.0
